Add Support for pango markup output
This format is recognized by xfce4-panel to allow plugins to display
pretty outputs. See the pango markup syntax here [1]
A plugin called xfce4-genmon-plugin can run a program periodically
and print its output in the panel.
A result similar to the one of the i3bar can be achieved easily by
having a properly configured panel.

Invoke i3status with --run-once and a different configuration file.
The current downside of this approach is the performance penalty
incured by executing the binary at each refresh.

[1] https://developer.gnome.org/pango/unstable/PangoMarkupFormat.html